+++ lib/listModule.js
@@ -1,12 +1,11 @@
 'use strict'
 
-const {assert} = require('chai')
+const listProductPopularity = function() {
 
-describe(`List Module`, function() {
-	describe(`listProductPopularity`, function() {
+}
 
-	})
-	describe(`listProducts`, function() {
+const listProducts = function() {
 
-	})
-})
+}
+
+module.exports = {listProducts, listProductPopularity}
